+++ src/unfold/static/unfold/js/app.js
@@ -14,11 +14,37 @@ window.addEventListener("load", () => {
 	tabNavigation();
 
 	scrollSidebarNav();
+
+	crispyFormset();
 });
+
+function getCurrentTab() {
+	const fragment = window.location.hash?.replace("#", "");
+
+	if (!fragment) {
+		return null;
+	}
+
+	if (!document.getElementById(`${fragment}-group`)) {
+		return null;
+	}
+
+	return fragment;
+}
 
 /*************************************************************
  * Switch theme
  *************************************************************/
+function isFilterModalOpen() {
+	const changelistFilter = document.getElementById("changelist-filter-close");
+
+	if (changelistFilter) {
+		return window.getComputedStyle(changelistFilter).display === "block";
+	}
+
+	return false;
+}
+
 function theme(defaultTheme = "auto") {
 	return {
 		sidebarWidth: localStorage.getItem("sidebarWidth") || 288,
@@ -43,6 +69,7 @@ function theme(defaultTheme = "auto") {
 		},
 		openModal: false,
 		filterOpen: false,
+		filterModalOpen: false,
 		openAllApplications: false,
 		adminTheme: Alpine.$persist(defaultTheme).as("adminTheme"),
 		init() {
@@ -59,14 +86,18 @@ function theme(defaultTheme = "auto") {
 			});
 
 			this.$watch("filterOpen", (value) => {
-				if (value) {
-					document
-						.getElementsByTagName("body")[0]
-						.classList.add("overflow-hidden");
-				} else {
-					document
-						.getElementsByTagName("body")[0]
-						.classList.remove("overflow-hidden");
+				if (isFilterModalOpen()) {
+					if (value) {
+						this.filterModalOpen = true;
+						document
+							.getElementsByTagName("body")[0]
+							.classList.add("overflow-hidden");
+					} else {
+						this.filterModalOpen = false;
+						document
+							.getElementsByTagName("body")[0]
+							.classList.remove("overflow-hidden");
+					}
 				}
 			});
 
@@ -110,16 +141,36 @@ function theme(defaultTheme = "auto") {
 				return "";
 			},
 			["x-on:keydown.window"](event) {
-				if (
-					event.key === "[" &&
-					!event.metaKey &&
-					!event.ctrlKey &&
-					document.activeElement.tagName.toLowerCase() !== "input" &&
-					document.activeElement.tagName.toLowerCase() !== "textarea" &&
-					!document.activeElement.isContentEditable
-				) {
+				const isInput =
+					document.activeElement.tagName.toLowerCase() === "input" ||
+					document.activeElement.tagName.toLowerCase() === "textarea" ||
+					document.activeElement.isContentEditable;
+
+				if (isInput) {
+					return;
+				}
+
+				if (!event.metaKey && !event.ctrlKey && event.key === "[") {
 					event.preventDefault();
 					this.sidebarToggle();
+				}
+
+				if (!event.metaKey && !event.ctrlKey && event.key === "f") {
+					const filterOpenButton = document.querySelector(
+						".filter-open-button",
+					);
+
+					if (filterOpenButton) {
+						this.filterOpen = !this.filterOpen;
+					}
+				}
+
+				if (!event.metaKey && !event.ctrlKey && event.key === "c") {
+					const addLink = document.querySelector(".addlink");
+
+					if (addLink) {
+						addLink.click();
+					}
 				}
 
 				if ((event.metaKey || event.ctrlKey) && event.key === "e") {
@@ -146,10 +197,9 @@ function scrollSidebarNav() {
 		return;
 	}
 
-	const instance = SimpleBar.instances.get(sidebarNav);
 	const activeItem = sidebarNav.querySelector("a.active");
 
-	if (!instance || !activeItem) {
+	if (!activeItem) {
 		return;
 	}
 
@@ -162,8 +212,10 @@ function scrollSidebarNav() {
 		);
 	}
 
-	if (instance && !isActiveItemVisible()) {
-		instance.getScrollElement().scroll(0, activeItem.offsetTop);
+	if (!isActiveItemVisible()) {
+		sidebarNav.scrollTo({
+			top: activeItem.offsetTop,
+		});
 	}
 }
 
@@ -551,9 +603,10 @@ function dateTimeShortcutsOverlay() {
 	const observer = new MutationObserver((mutations) => {
 		for (const mutationRecord of mutations) {
 			const display = mutationRecord.target.style.display;
+			const hasOpenAttribute = mutationRecord.target.hasAttribute("open");
 			const overlay = document.getElementById("modal-overlay");
 
-			if (display === "block") {
+			if (display === "block" || hasOpenAttribute) {
 				overlay.style.display = "block";
 			} else {
 				overlay.style.display = "none";
@@ -565,7 +618,7 @@ function dateTimeShortcutsOverlay() {
 		for (const target of document.querySelectorAll(".calendarbox, .clockbox")) {
 			observer.observe(target, {
 				attributes: true,
-				attributeFilter: ["style"],
+				attributeFilter: ["style", "open"],
 			});
 		}
 	};
@@ -1106,16 +1159,125 @@ document.addEventListener("htmx:afterSettle", (e) => {
 	});
 });
 
-function getCurrentTab() {
-	const fragment = window.location.hash?.replace("#", "");
+/*************************************************************
+ * Crispy formset
+ *************************************************************/
+function crispyFormsetReindex(formsetId) {
+	document
+		.querySelectorAll(`#${formsetId} .dynamic-form`)
+		.forEach((row, index) => {
+			const newIndex =
+				document.querySelectorAll(`#${formsetId} .original-form`).length +
+				index;
 
-	if (!fragment) {
-		return null;
+			["id", "name", "for"].forEach((attr) => {
+				row.querySelectorAll(`[${attr}]`).forEach((el) => {
+					const val = el.getAttribute(attr);
+
+					if (val) {
+						el.setAttribute(attr, val.replace(/-\d+-/g, `-${newIndex}-`));
+					}
+				});
+			});
+		});
+}
+
+function crispyFormsetDelete(target) {
+	const rows = target
+		? target.querySelectorAll(".crispy-formset-delete")
+		: document.querySelectorAll(".crispy-formset-delete");
+
+	rows.forEach((el) => {
+		el.addEventListener("click", () => {
+			const formsetId = el.dataset.formsetId;
+			const formTotalEl = document.querySelector(
+				`#${formsetId} input[name*="TOTAL_FORMS"]`,
+			);
+			const formMinEl = document.querySelector(
+				`#${formsetId} input[name*="MIN_NUM_FORMS"]`,
+			);
+			const totalForms = parseInt(formTotalEl.value, 10);
+			const minNumForms = parseInt(formMinEl.value, 10);
+
+			if (totalForms <= minNumForms) {
+				return;
+			}
+
+			const rowToDelete = el.closest(".dynamic-form");
+
+			rowToDelete.remove();
+			formTotalEl.value = formTotalEl.value - 1;
+
+			crispyFormsetReindex(formsetId);
+			crispyFormsetToggleAdd(formsetId);
+		});
+	});
+}
+
+function crispyFormsetAdd(target) {
+	const formsetId = target.dataset.formsetId;
+	const formTotalEl = document.querySelector(
+		`#${formsetId} input[name*="TOTAL_FORMS"]`,
+	);
+	const formCount = parseInt(formTotalEl.value, 10);
+
+	const newForm = document
+		.querySelector(`#${formsetId} .empty-form`)
+		.cloneNode(true);
+
+	newForm.classList.remove("empty-form", "hidden");
+	newForm.classList.add("dynamic-form");
+	newForm.innerHTML = newForm.innerHTML.replaceAll(/__prefix__/g, formCount);
+
+	document.getElementById(`${formsetId}-rows`).insertBefore(newForm, null);
+
+	formTotalEl.value = formCount + 1;
+
+	newForm.dispatchEvent(
+		new CustomEvent("formset:added", {
+			bubbles: true,
+		}),
+	);
+
+	crispyFormsetToggleAdd(formsetId);
+}
+
+function crispyFormsetToggleAdd(formsetId) {
+	const totalForms = parseInt(
+		document.querySelector(`#${formsetId} input[name*="TOTAL_FORMS"]`).value,
+		10,
+	);
+
+	const maxNumForms = parseInt(
+		document.querySelector(`#${formsetId} input[name*="MAX_NUM_FORMS"]`).value,
+		10,
+	);
+
+	if (totalForms >= maxNumForms) {
+		document
+			.getElementById(`${formsetId}-add-row-wrapper`)
+			.classList.add("hidden");
+	} else {
+		document
+			.getElementById(`${formsetId}-add-row-wrapper`)
+			.classList.remove("hidden");
 	}
+}
 
-	if (!document.getElementById(`${fragment}-group`)) {
-		return null;
-	}
+function crispyFormset() {
+	document
+		.querySelectorAll(".crispy-formset-add")
+		.forEach((formsetAddButton) => {
+			crispyFormsetToggleAdd(formsetAddButton.dataset.formsetId);
 
-	return fragment;
+			formsetAddButton.addEventListener("click", () => {
+				crispyFormsetAdd(formsetAddButton);
+			});
+		});
+
+	crispyFormsetDelete();
+
+	document.addEventListener("formset:added", (event) => {
+		crispyFormsetDelete(event.target);
+	});
 }