+++ CHANGELOG.md
@@ -1,6 +1,106 @@
 # CHANGELOG
 
 
+## v0.101.0 (2026-07-17)
+
+### Bug Fixes
+
+- Choice filter dark background color
+  ([#2148](https://github.com/unfoldadmin/django-unfold/pull/2148),
+  [`db49a7d`](https://github.com/unfoldadmin/django-unfold/commit/db49a7d64d9539e934c08900ddedb6fd70f9011e))
+
+- Firefox background rendering ([#2151](https://github.com/unfoldadmin/django-unfold/pull/2151),
+  [`b914d3d`](https://github.com/unfoldadmin/django-unfold/commit/b914d3da9a25096bb07f23edcea0c309f6fcfda9))
+
+- Optional sidebar scrollbar ([#2152](https://github.com/unfoldadmin/django-unfold/pull/2152),
+  [`647ee27`](https://github.com/unfoldadmin/django-unfold/commit/647ee2709952d5817610a3ef2919f8139a4fa6b9))
+
+- Preserve filters after delete (#1897)
+  ([#2149](https://github.com/unfoldadmin/django-unfold/pull/2149),
+  [`a90fc6f`](https://github.com/unfoldadmin/django-unfold/commit/a90fc6ff8ec4296999c82b276226700916024f1d))
+
+- Table scrollbar ([#2153](https://github.com/unfoldadmin/django-unfold/pull/2153),
+  [`95c4f68`](https://github.com/unfoldadmin/django-unfold/commit/95c4f683e3625960f35c42a7d5c61c6d60ea6540))
+
+- Trap focus for command ([#2146](https://github.com/unfoldadmin/django-unfold/pull/2146),
+  [`1ce2af2`](https://github.com/unfoldadmin/django-unfold/commit/1ce2af25936b4721b58d4feddb1dba93d9d9dd1f))
+
+- Trap modal focus ([#2147](https://github.com/unfoldadmin/django-unfold/pull/2147),
+  [`02ae1de`](https://github.com/unfoldadmin/django-unfold/commit/02ae1de92194a8f93509814d307629e4d7d5fcc6))
+
+### Features
+
+- Simplify scrollbars ([#2150](https://github.com/unfoldadmin/django-unfold/pull/2150),
+  [`fc2eaa9`](https://github.com/unfoldadmin/django-unfold/commit/fc2eaa9b2622d1ca22e1003d760314168e2e5814))
+
+- Table component css class ([#2138](https://github.com/unfoldadmin/django-unfold/pull/2138),
+  [`4c5f51f`](https://github.com/unfoldadmin/django-unfold/commit/4c5f51f66475fe979b441ef8dc572888c067da76))
+
+- Table component scrolling ([#2139](https://github.com/unfoldadmin/django-unfold/pull/2139),
+  [`b3fe775`](https://github.com/unfoldadmin/django-unfold/commit/b3fe7755b26547b30c552d9a86d2ab269f1c3f5a))
+
+- Trap focus for modal filter ([#2145](https://github.com/unfoldadmin/django-unfold/pull/2145),
+  [`a71d2ec`](https://github.com/unfoldadmin/django-unfold/commit/a71d2ec76f19e7d1b75015ff06f2afb2afe8ea45))
+
+
+## v0.100.0 (2026-07-06)
+
+### Bug Fixes
+
+- Badge callback ([#2133](https://github.com/unfoldadmin/django-unfold/pull/2133),
+  [`916e045`](https://github.com/unfoldadmin/django-unfold/commit/916e04579daffb10c2be0c1fa12a0561e95a746b))
+
+- Filter background color ([#2132](https://github.com/unfoldadmin/django-unfold/pull/2132),
+  [`5b6c4d7`](https://github.com/unfoldadmin/django-unfold/commit/5b6c4d753a1e258c954a0a42bd83c5addd153034))
+
+### Documentation
+
+- Crispy form doc demo URL ([#2135](https://github.com/unfoldadmin/django-unfold/pull/2135),
+  [`fa014e8`](https://github.com/unfoldadmin/django-unfold/commit/fa014e83e54825be491fc7471e310b49fa17ad30))
+
+- Custom site app name ([#2047](https://github.com/unfoldadmin/django-unfold/pull/2047),
+  [`d79567c`](https://github.com/unfoldadmin/django-unfold/commit/d79567cd78c4e3abb62c652cda8588977fe13b1c))
+
+### Features
+
+- Add link shortcut ([#2130](https://github.com/unfoldadmin/django-unfold/pull/2130),
+  [`19de972`](https://github.com/unfoldadmin/django-unfold/commit/19de9721fbbb1b604b668eef88ff7b01be3f0e66))
+
+- Toggle filter shortcut ([#2131](https://github.com/unfoldadmin/django-unfold/pull/2131),
+  [`675fdb1`](https://github.com/unfoldadmin/django-unfold/commit/675fdb1017060cc8ff2cd683b8e03186b0785e13))
+
+
+## v0.99.1 (2026-07-01)
+
+### Bug Fixes
+
+- Django 6.1 calendar and time ([#2118](https://github.com/unfoldadmin/django-unfold/pull/2118),
+  [`929ee15`](https://github.com/unfoldadmin/django-unfold/commit/929ee15c8fc9c9af42eec0f33640b5016ed210a1))
+
+- Internal dataset name ([#2128](https://github.com/unfoldadmin/django-unfold/pull/2128),
+  [`51f41ea`](https://github.com/unfoldadmin/django-unfold/commit/51f41eaa0db840171f8f29644b40749cd4c1b059))
+
+### Documentation
+
+- Typo in table component
+  ([`8b50da0`](https://github.com/unfoldadmin/django-unfold/commit/8b50da0b2ef98e7966c1f25bbe7277e95c0408c2))
+
+
+## v0.99.0 (2026-06-23)
+
+### Features
+
+- Crispy formset delete rows ([#2112](https://github.com/unfoldadmin/django-unfold/pull/2112),
+  [`ad8c6dd`](https://github.com/unfoldadmin/django-unfold/commit/ad8c6dd9b1a33db3eff001e9fe95891b8b4a6cd1))
+
+- Crispy formset respect min/max num of forms
+  ([#2115](https://github.com/unfoldadmin/django-unfold/pull/2115),
+  [`e876810`](https://github.com/unfoldadmin/django-unfold/commit/e876810c21616120684f9d971b1890b3c013bb87))
+
+- Custom site views ([#2022](https://github.com/unfoldadmin/django-unfold/pull/2022),
+  [`7a9ef75`](https://github.com/unfoldadmin/django-unfold/commit/7a9ef75f0582bd68252e417864efebf7cfab1112))
+
+
 ## v0.98.0 (2026-06-19)
 
 ### Bug Fixes